remove redundant cloudbuild steps
In the cloud funcitons module we had various python an terraform
static analysis steps that were redundant with the repo wide
pre-commit which will already be run by CI triggers. I've trimmed
it to include only the tests specific to the cloud function and
removed the duplicate static analysis configs in favor of the repo
wide configs.