Update job_execution.sql (#451)
Changed the calculation of job_duration_seconds in the job_execution.sql

If the job is running then end_time is null. Therefore an admin cannot easily visualize for how long the a job has been running until it is done.

With this change, if the job is not done then job_duration_seconds is calculates as the difference between start time and current time. Thus and admin can know easily identify the duration of a RUNNING / PENDING job and identify long running / troublesome jobs.